// server/src/types/generated.d.ts
export type Maybe<T> = T | null;
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};









export type AdditionalEntityFields = {
  path?: Maybe<Scalars['String']>;
  type?: Maybe<Scalars['String']>;
};

export type AuthInput = {
  email: Scalars['String'];
  password: Scalars['String'];
};

export type Mutation = {
   __typename?: 'Mutation';
  signup?: Maybe<Scalars['String']>;
  updateProfile?: Maybe<UserType>;
  updateEmail?: Maybe<UserType>;
  updatePassword?: Maybe<UserType>;
  deleteUser?: Maybe<Scalars['String']>;
};


export type MutationSignupArgs = {
  input: AuthInput;
};


export type MutationUpdateProfileArgs = {
  input: UpdateProfileInput;
};


export type MutationUpdateEmailArgs = {
  input?: Maybe<Scalars['String']>;
};


export type MutationUpdatePasswordArgs = {
  input?: Maybe<Scalars['String']>;
};

export type Query = {
   __typename?: 'Query';
  login?: Maybe<Scalars['String']>;
  getUser?: Maybe<UserType>;
};


export type QueryLoginArgs = {
  input: AuthInput;
};

export type UpdateProfileInput = {
  firstName?: Maybe<Scalars['String']>;
  lastName?: Maybe<Scalars['String']>;
};

export type UserProfileType = {
   __typename?: 'UserProfileType';
  firstName?: Maybe<Scalars['String']>;
  lastName?: Maybe<Scalars['String']>;
};

export type UserType = {
   __typename?: 'UserType';
  _id: Scalars['ID'];
  email: Scalars['String'];
  createdAt: Scalars['String'];
  updatedAt: Scalars['String'];
  profile?: Maybe<UserProfileType>;
};




export type ResolverTypeWrapper<T> = Promise<T> | T;


export type StitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;
// ...
export type isTypeOfResolverFn<T = {}> = (obj: T, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Query: ResolverTypeWrapper<{}>,
  AuthInput: AuthInput,
  String: ResolverTypeWrapper<Scalars['String']>,
  UserType: ResolverTypeWrapper<UserType>,
  ID: ResolverTypeWrapper<Scalars['ID']>,
  UserProfileType: ResolverTypeWrapper<UserProfileType>,
  Mutation: ResolverTypeWrapper<{}>,
  UpdateProfileInput: UpdateProfileInput,
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>,
  AdditionalEntityFields: AdditionalEntityFields,
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Query: {},
  AuthInput: AuthInput,
  String: Scalars['String'],
  UserType: UserType,
  ID: Scalars['ID'],
  UserProfileType: UserProfileType,
  Mutation: {},
  UpdateProfileInput: UpdateProfileInput,
  Boolean: Scalars['Boolean'],
  AdditionalEntityFields: AdditionalEntityFields,
};
// ...
export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  signup?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<MutationSignupArgs, 'input'>>,
  updateProfile?: Resolver<Maybe<ResolversTypes['UserType']>, ParentType, ContextType, RequireFields<MutationUpdateProfileArgs, 'input'>>,
  updateEmail?: Resolver<Maybe<ResolversTypes['UserType']>, ParentType, ContextType, RequireFields<MutationUpdateEmailArgs, never>>,
  updatePassword?: Resolver<Maybe<ResolversTypes['UserType']>, ParentType, ContextType, RequireFields<MutationUpdatePasswordArgs, never>>,
  deleteUser?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>,
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  login?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<QueryLoginArgs, 'input'>>,
  getUser?: Resolver<Maybe<ResolversTypes['UserType']>, ParentType, ContextType>,
};

export type UserProfileTypeResolvers<ContextType = any, ParentType extends ResolversParentTypes['UserProfileType'] = ResolversParentTypes['UserProfileType']> = {
  firstName?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>,
  lastName?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>,
  __isTypeOf?: isTypeOfResolverFn<ParentType>,
};

export type UserTypeResolvers<ContextType = any, ParentType extends ResolversParentTypes['UserType'] = ResolversParentTypes['UserType']> = {
  _id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>,
  email?: Resolver<ResolversTypes['String'], ParentType, ContextType>,
  createdAt?: Resolver<ResolversTypes['String'], ParentType, ContextType>,
  updatedAt?: Resolver<ResolversTypes['String'], ParentType, ContextType>,
  profile?: Resolver<Maybe<ResolversTypes['UserProfileType']>, ParentType, ContextType>,
  __isTypeOf?: isTypeOfResolverFn<ParentType>,
};

export type Resolvers<ContextType = any> = {
  Mutation?: MutationResolvers<ContextType>,
  Query?: QueryResolvers<ContextType>,
  UserProfileType?: UserProfileTypeResolvers<ContextType>,
  UserType?: UserTypeResolvers<ContextType>,
};